refactor(academy): move the gate notice into the widgets layer (AYC-253)
- features/academy now carries only hooks and pure logic — its ui/
  directory is gone; AcademyGateNotice becomes widgets/academy-gate-notice,
  matching app-alert-banner and certificate-expiry
- the three consuming pages import the widget; feature public API shrinks
