+++ ayunis-core-frontend/src/features/SUMMARY.md
@@ -4,4 +4,4 @@ Self-contained business logic modules shared across multiple pages and widgets.
 
 Eight feature modules and several standalone hooks provide cross-cutting concerns: theme management (dark/light mode persistence), language switching (German/English via i18next), model/provider configuration with provider icon utilities, usage analytics data hooks for statistics, charts, and distribution metrics, skill action hooks for toggling skill active/pinned state, marketplace configuration for exposing marketplace availability and destination URLs, fair use limit editing for shared limit validation and edit state, and permission gating for hiding UI controls the current user lacks access to.
 
-The `theme/` feature manages dark/light mode with localStorage persistence through `useTheme`, `setTheme`, and `getTheme` utilities. The `language/` feature wraps i18next language switching in a `useLanguage` hook. The `models/` feature provides `useLanguageModels`, `useEmbeddingModels`, `useImageGenerationModels`, and `useProviders` hooks for fetching AI model configurations by type, plus provider constants and icon utilities for Ollama, Mistral, Anthropic, and OpenAI. The `usage/` feature contains analytics hooks (`useUsageStats`, `useUserUsage`, `useProviderUsage`, `useProviderUsageChart`, `useModelDistribution`, `useUsageConfig`) powering admin dashboard charts. The `skill-actions/` feature provides `useToggleSkillActive` and `useToggleSkillPinned` hooks for toggling skill active and pinned state. The `marketplace/` feature provides a `useMarketplaceConfig` hook that returns marketplace availability plus the configured destination URL for marketplace entry points. The `fair-use-limit-editor/` feature provides `useFairUseLimitEditor` for reusable fair use limit edit state and validation. The `academy/` feature exposes the academy add-on gate (`isAcademyAddonActive`, `useIsAcademyAddonActive`), `useDownloadCertificate` for downloading the KI-Führerschein certificate PDF (used by the quiz result card and the academy overview completion banner), `useAcademyAccessStatus` plus the `AcademyGateNotice` component for the org-level certificate requirement that gates Core chat, and `useAcademyProgress` for per-chapter pass state and the whole-academy completion date — shared by the academy overview and the account-settings certificate card, and add-on gated server-side, so callers outside the academy pass `enabled: false` while the add-on is inactive. For orgs on annual renewal it also owns the pre-expiry warnings: the pure `resolveCertificateExpiryNotice` maps days-until-expiry to a notice level (dismissible pop-up at 30 days, red pop-up at 14, non-dismissible countdown banner in the final 7); the components that render those notices live in `widgets/certificate-expiry`, so this feature stays hooks and pure logic. The `permissions/` feature provides `useMyPermissions` (with a `can(permission)` helper, backed by `GET /permissions/me`) for gating UI controls on the current user's effective RBAC permissions — e.g. hiding the create-skill / create-knowledge-base controls from members who lack the permission — plus the `PermissionGate` component (`ui/`) that renders its children only for holders of any of the given permissions, and `allowedSettingsSections(role, permissions)` with the `MANAGER_SETTINGS_SECTIONS` map (`lib/`), the single answer to "may this user open this admin-settings section?" used by the `admin-settings` route guard and both sidebars (admins get the settings root, which prefixes every section). Standalone hooks at the feature root include `useAgents` for agent listing, `useDeleteThread` and `useRenameThread` for thread management, `useAutoScroll` for chat scrolling behavior, `useContentScrollHeader` for tracking whether scrolled content sits under the frosted overlay header, `useChatsSidebarOpen` for persisting chats sidebar collapsed state, `useKeyboardShortcut` for hotkey bindings, `usePermittedModels` for org-scoped model access, `useReleaseNotes` for changelog display, `useRedirectNotification` for post-navigation toasts, and `useGtm` for Google Tag Manager integration.
+The `theme/` feature manages dark/light mode with localStorage persistence through `useTheme`, `setTheme`, and `getTheme` utilities. The `language/` feature wraps i18next language switching in a `useLanguage` hook. The `models/` feature provides `useLanguageModels`, `useEmbeddingModels`, `useImageGenerationModels`, and `useProviders` hooks for fetching AI model configurations by type, plus provider constants and icon utilities for Ollama, Mistral, Anthropic, and OpenAI. The `usage/` feature contains analytics hooks (`useUsageStats`, `useUserUsage`, `useProviderUsage`, `useProviderUsageChart`, `useModelDistribution`, `useUsageConfig`) powering admin dashboard charts. The `skill-actions/` feature provides `useToggleSkillActive` and `useToggleSkillPinned` hooks for toggling skill active and pinned state. The `marketplace/` feature provides a `useMarketplaceConfig` hook that returns marketplace availability plus the configured destination URL for marketplace entry points. The `fair-use-limit-editor/` feature provides `useFairUseLimitEditor` for reusable fair use limit edit state and validation. The `academy/` feature exposes the academy add-on gate (`isAcademyAddonActive`, `useIsAcademyAddonActive`), `useDownloadCertificate` for downloading the KI-Führerschein certificate PDF (used by the quiz result card and the academy overview completion banner), `useAcademyAccessStatus` for the org-level certificate requirement that gates Core chat (the notice explaining the gate lives in `widgets/academy-gate-notice`), and `useAcademyProgress` for per-chapter pass state and the whole-academy completion date — shared by the academy overview and the account-settings certificate card, and add-on gated server-side, so callers outside the academy pass `enabled: false` while the add-on is inactive. For orgs on annual renewal it also owns the pre-expiry warnings: the pure `resolveCertificateExpiryNotice` maps days-until-expiry to a notice level (dismissible pop-up at 30 days, red pop-up at 14, non-dismissible countdown banner in the final 7); the components that render those notices live in `widgets/certificate-expiry`, so this feature stays hooks and pure logic. The `permissions/` feature provides `useMyPermissions` (with a `can(permission)` helper, backed by `GET /permissions/me`) for gating UI controls on the current user's effective RBAC permissions — e.g. hiding the create-skill / create-knowledge-base controls from members who lack the permission — plus the `PermissionGate` component (`ui/`) that renders its children only for holders of any of the given permissions, and `allowedSettingsSections(role, permissions)` with the `MANAGER_SETTINGS_SECTIONS` map (`lib/`), the single answer to "may this user open this admin-settings section?" used by the `admin-settings` route guard and both sidebars (admins get the settings root, which prefixes every section). Standalone hooks at the feature root include `useAgents` for agent listing, `useDeleteThread` and `useRenameThread` for thread management, `useAutoScroll` for chat scrolling behavior, `useContentScrollHeader` for tracking whether scrolled content sits under the frosted overlay header, `useChatsSidebarOpen` for persisting chats sidebar collapsed state, `useKeyboardShortcut` for hotkey bindings, `usePermittedModels` for org-scoped model access, `useReleaseNotes` for changelog display, `useRedirectNotification` for post-navigation toasts, and `useGtm` for Google Tag Manager integration.
