+++ ayunis-core-frontend/src/features/academy/index.ts
@@ -3,7 +3,6 @@ export { useIsAcademyAddonActive } from './useIsAcademyAddonActive';
 export { useDownloadCertificate } from './useDownloadCertificate';
 export { useAcademyAccessStatus } from './useAcademyAccessStatus';
 export { useAcademyProgress } from './useAcademyProgress';
-export { default as AcademyGateNotice } from './ui/AcademyGateNotice';
 export { resolveCertificateExpiryNotice } from './lib/resolve-certificate-expiry-notice';
 export type { CertificateExpiryNoticeLevel } from './lib/resolve-certificate-expiry-notice';
 export { ACADEMY_LANDING_PAGE_URL } from './constants';
