+++ ayunis-core-frontend/src/widgets/SUMMARY.md
@@ -2,6 +2,6 @@
 
 Composite UI components reused across two or more application pages.
 
-Widget modules provide shared interactive components: app sidebar with thread navigation and user menu, chat input with voice recording and image support, artifact editing, the spreadsheet-editor widget with an editable grid, version history, and XLSX/CSV export, markdown rendering with syntax-highlighted code blocks, chart visualizations (line, bar, pie), MCP user-configuration forms, confirmation modal with context provider, a one-time welcome-video dialog, the certificate-expiry notices (pre-expiry warning dialog plus the final-week countdown banner, both mounted in `AppLayout`; dismissal state lives in the slice's `model/` hook and the notice thresholds come from `features/academy`), and utility widgets for headers, empty states, tooltips, and thread renaming.
+Widget modules provide shared interactive components: app sidebar with thread navigation and user menu, chat input with voice recording and image support, artifact editing, the spreadsheet-editor widget with an editable grid, version history, and XLSX/CSV export, markdown rendering with syntax-highlighted code blocks, chart visualizations (line, bar, pie), MCP user-configuration forms, confirmation modal with context provider, a one-time welcome-video dialog, the academy gate notice (why chat is locked, shown by the chat, new-chat and academy pages), the certificate-expiry notices (pre-expiry warning dialog plus the final-week countdown banner, both mounted in `AppLayout`; dismissal state lives in the slice's `model/` hook and the notice thresholds come from `features/academy`), and utility widgets for headers, empty states, tooltips, and thread renaming.
 
 The `app-sidebar/` widget is the primary navigation component featuring chat thread listing, logout functionality, subscription status hints, release notes access, and user profile display—powered by dedicated API hooks (`useThreads`, `useLogout`, `useMe`, `useHasActiveSubscription`). The `chat-input/` widget handles message composition with rich features: voice recording via `useVoiceRecording`, file drag-and-drop via `useFileDrop`, image paste support via `useImagePaste`, pending image management, audio transcription through `useTranscribe`, and knowledge base attachment via `useKnowledgeBases`. The knowledge base integration adds a "Knowledge Bases" section to the `PlusButton` dropdown (with loading, error, empty, and all-attached states), renders attached knowledge bases as removable chips in `SourcesList`, and passes pending KB state (`KnowledgeBaseSummary[]`) through `ChatInput` props (`onAddKnowledgeBase`, `onRemoveKnowledgeBase`). The `artifact-editor/` widget provides the chat side-panel WYSIWYG document editor with save, revert, DOCX/PDF export, and a `LetterheadPicker` for switching the artifact's assigned organization letterhead. The `markdown/` widget renders AI responses with a custom `Codeblock` component for syntax highlighting. The `charts/` widget group provides `LineChartWidget`, `BarChartWidget`, and `PieChartWidget` wrapped in `ChartCard` containers with loading and empty states, used in admin usage dashboards. The `confirmation-modal/` widget uses React context (`ConfirmationProvider`) for imperative confirmation dialogs via `useConfirmation`. Supporting widgets include `content-area-header/`, `empty-state/`, `tooltip-if/` (conditional tooltips), `agent-activity-hint/`, and `rename-thread-dialog/`. The `model-type-card/` widget displays language or embedding model lists with enable/disable toggles and anonymous-only configuration, shared between admin and super-admin model settings pages; it also exports `ModelTierStars` and `TierStars` for the model catalog and `model-select-options/`, with translated performance and usage labels and no stars for untiered models. The `model-select-options/` widget renders the shared model selector options and their hover info cards, including catalog-managed descriptions, tier indicators, and provider-hosting details. The `subscription-cancellation-section/` widget handles subscription cancellation and reactivation with confirmation dialogs, shared between admin and super-admin subscription pages. Additional analytics widgets include `model-distribution-chart/`, `provider-consumption-chart/`, `usage-stats-cards/`, `user-usage-table/`, and `shares-tab/`. The `credit-budget-display/` widget provides a reusable credit budget visualization with loading, error, and data states—displaying monthly budget, used credits, remaining credits, and a usage progress bar—shared between the admin usage page and super-admin organization page. The `margin-editor/` widget renders a PDF page preview with an interactive margin overlay editor, used in both the letterhead creation dialog and the letterhead detail page. The `onboarding/` widget contains the getting-started tour engine and progress model: `OnboardingTourProvider` (mounted in the authenticated layout, rendering a lazy react-joyride `TourRenderer`), `OnboardingTourTarget` (a `data-tour` wrapper that pages and the chat-input widget use to mark spotlight targets from the `TOUR_TARGET` registry), `OnboardingReturnButton` (floating return-to-onboarding button), the `useOnboarding` progress query, the onboarding step catalog (`config/categories.ts`), and the `useOnboardingProgress` derivation hook consumed by the app-sidebar's `OnboardingCard`; the `/getting-started` page in `pages/onboarding` composes these. The `mfa-enrollment/` widget provides the shared two-factor enrollment UI: `MfaEnrollmentPanel` (QR code plus manual-entry secret with a six-digit TOTP confirmation form) and `RecoveryCodesPanel` (one-time recovery code display with copy-all, gated behind an explicit continue button), shared between the forced-enrollment step of the two-factor login page and the account-settings two-factor card.
