+++ ayunis-core-frontend/src/widgets/academy-gate-notice/index.ts
@@ -1,0 +1,1 @@
+export { default as AcademyGateNotice } from './ui/AcademyGateNotice';
